A&GのHLS対応

+++ internal/recorder/recorder.go
@@ -73,18 +73,20 @@ func Record(r Recorder) error {
 		out, cmdErr := exec.Command(recC[0], recC[1:]...).Output()
 		fmt.Println(out, cmdErr)
 
-		// start converting
-		fmt.Println("Converting...")
-		convC, convErr := shellwords.Parse(mp4Cmd)
-		if convErr != nil {
-			return
-		}
-		convO, convE := exec.Command(convC[0], convC[1:]...).Output()
-		fmt.Println(convO, convE)
+		if _, err = os.Stat(fileManager.OutputPath + ".flv"); err == nil {
+			// start converting
+			fmt.Println("Converting...")
+			convC, convErr := shellwords.Parse(mp4Cmd)
+			if convErr != nil {
+				return
+			}
+			convO, convE := exec.Command(convC[0], convC[1:]...).Output()
+			fmt.Println(convO, convE)
 
-		// remove src flv file
-		if rmErr := os.Remove(fileManager.OutputPath + ".flv"); rmErr != nil {
-			return
+			// remove src flv file
+			if rmErr := os.Remove(fileManager.OutputPath + ".flv"); rmErr != nil {
+				return
+			}
 		}
 
 		// register data to table
